refactor: use fastify (#661)
