+++ package.json
@@ -81,6 +81,8 @@
     "@emotion/babel-plugin": "^11.11.0",
     "@emotion/css": "^11.11.0",
     "@emotion/react": "^11.11.1",
+    "@fastify/express": "^2.3.0",
+    "@fastify/static": "^6.10.2",
     "@lingui/loader": "^3.17.2",
     "@lingui/react": "^3.17.2",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
@@ -92,7 +94,6 @@
     "@types/cookie-parser": "^1.4.3",
     "@types/create-torrent": "^5.0.0",
     "@types/d3": "^7.4.0",
-    "@types/debug": "^4.1.8",
     "@types/express": "^4.17.17",
     "@types/fs-extra": "^9.0.13",
     "@types/geoip-country": "^4.0.0",
@@ -104,7 +105,7 @@
     "@types/node": "^12.20.55",
     "@types/parse-torrent": "^5.8.4",
     "@types/passport": "^1.0.12",
-    "@types/passport-jwt": "^3.0.8",
+    "@types/passport-jwt": "^3.0.9",
     "@types/react": "^18.2.11",
     "@types/react-dom": "^18.2.4",
     "@types/react-measure": "^2.0.8",
@@ -136,8 +137,6 @@
     "d3-scale": "^4.0.2",
     "d3-selection": "^3.0.0",
     "d3-shape": "^3.2.0",
-    "debug": "^4.3.4",
-    "esbuild-jest": "^0.5.0",
     "eslint": "^8.42.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.0.0",
@@ -149,6 +148,8 @@
     "express-rate-limit": "^6.7.0",
     "fast-json-patch": "^3.1.1",
     "fast-sort": "^3.4.0",
+    "fastify": "^4.21.0",
+    "fastify-type-provider-zod": "^1.1.9",
     "feedsub": "^0.7.8",
     "file-loader": "^6.2.0",
     "form-data": "^4.0.0",
@@ -159,6 +160,7 @@
     "html-webpack-plugin": "^5.5.3",
     "http-errors": "^2.0.0",
     "jest": "^28.1.3",
+    "jest-esbuild": "^0.3.0",
     "js-file-download": "^0.4.12",
     "jsonwebtoken": "^9.0.0",
     "lodash": "^4.17.21",
@@ -176,7 +178,6 @@
     "postcss": "^8.4.24",
     "postcss-loader": "^7.3.3",
     "prettier": "^2.8.8",
-    "promise": "^8.3.0",
     "react": "^18.2.0",
     "react-dev-utils": "^12.0.1",
     "react-dom": "^18.2.0",
