+++ server/bin/start.ts
@@ -4,6 +4,7 @@ import chalk from 'chalk';
 
 import enforcePrerequisites from './enforce-prerequisites';
 import migrateData from './migrations/run';
+import startWebServer from './web-server';
 
 if (process.env.NODE_ENV == 'production') {
   // Catch unhandled rejections and exceptions
@@ -26,11 +27,7 @@ if (process.env.NODE_ENV == 'production') {
 
 enforcePrerequisites()
   .then(migrateData)
-  .then(() => {
-    // We do this because we don't want the side effects of importing server functions before migration is completed.
-    const startWebServer = require('./web-server').default; // eslint-disable-line @typescript-eslint/no-var-requires
-    return startWebServer();
-  })
+  .then(startWebServer)
   .catch((error) => {
     console.log(chalk.red('Failed to start Flood:'));
     console.trace(error);
