+++ server/bin/web-server.ts
@@ -1,106 +1,51 @@
 import chalk from 'chalk';
-import debug from 'debug';
+import fastify from 'fastify';
 import fs from 'fs';
-import http from 'http';
-import https from 'https';
 
-import app from '../app';
+import type {FastifyInstance} from 'fastify';
+import type {Http2SecureServer} from 'http2';
+import type {Server} from 'http';
+
 import config from '../../config';
+import constructRoutes from '../routes';
 import packageJSON from '../../package.json';
 
-const debugFloodServer = debug('flood:server');
+const startWebServer = async () => {
+  const {ssl = false, floodServerHost: host, floodServerPort: port} = config;
 
-// Normalize a port into a number, string, or false.
-const normalizePort = (val: string | number): string | number => {
-  const port = parseInt(val as string, 10);
+  let instance: FastifyInstance<Http2SecureServer> | FastifyInstance<Server>;
 
-  // Named pipe.
-  if (Number.isNaN(port)) {
-    return val;
-  }
-
-  // Port number.
-  if (port >= 0) {
-    return port;
-  }
-
-  console.error('Unexpected port or pipe');
-  process.exit(1);
-};
-
-const startWebServer = () => {
-  const port = normalizePort(config.floodServerPort);
-  const host = config.floodServerHost;
-  const useSSL = config.ssl ?? false;
-
-  app.set('port', port);
-  app.set('host', host);
-
-  // Create HTTP or HTTPS server.
-  let server: http.Server | https.Server;
-
-  if (useSSL) {
+  if (ssl) {
     if (!config.sslKey || !config.sslCert) {
       console.error('Cannot start HTTPS server, `sslKey` or `sslCert` is missing in config.js.');
       process.exit(1);
     }
 
-    server = https.createServer(
-      {
+    instance = fastify({
+      bodyLimit: 100 * 1024 * 1024,
+      trustProxy: 'loopback',
+      http2: true,
+      https: {
+        allowHTTP1: true,
         key: fs.readFileSync(config.sslKey),
         cert: fs.readFileSync(config.sslCert),
       },
-      app,
-    );
+    });
   } else {
-    server = http.createServer(app);
+    instance = fastify({
+      bodyLimit: 100 * 1024 * 1024,
+      trustProxy: 'loopback',
+    });
   }
 
-  const handleError = (error: NodeJS.ErrnoException) => {
-    if (error.syscall !== 'listen') {
-      throw error;
-    }
+  await constructRoutes(instance as FastifyInstance);
 
-    const bind = typeof port === 'string' ? `Pipe ${port}` : `Port ${port}`;
-
-    // Handle specific listen errors with friendly messages.
-    switch (error.code) {
-      case 'EACCES':
-        console.error(`${bind} requires elevated privileges`);
-        process.exit(1);
-      case 'EADDRINUSE':
-        console.error(`${bind} is already in use`);
-        process.exit(1);
-      default:
-        throw error;
-    }
-  };
-
-  // Event listener for HTTP server "listening" event.
-  const handleListening = () => {
-    const addr = server.address();
-    if (addr == null) {
-      console.error('Unable to get listening address.');
-      process.exit(1);
-    }
-    const bind = typeof addr === 'string' ? `pipe ${addr}` : `port ${addr.port}`;
-    debugFloodServer(`Listening on ${bind}`);
-  };
-
-  // Listen on provided port, on all network interfaces.
   if (typeof port === 'string') {
-    server.listen(port);
+    await instance.listen({path: port});
   } else {
-    server.listen(port, host);
+    await instance.listen({port, host});
   }
-
-  server.on('error', handleError);
-  server.on('listening', handleListening);
-  process.on('exit', () => {
-    server.close();
-  });
-
-  const address = chalk.underline(typeof port === 'string' ? port : `${useSSL ? 'https' : 'http'}://${host}:${port}`);
+  const address = chalk.underline(`${ssl ? 'https' : 'http'}://${host}:${port}`);
 
   console.log(chalk.green(`Flood server ${packageJSON.version} starting on ${address}\n`));
