+++ server/routes/api/index.test.ts
@@ -1,11 +1,23 @@
+import fastify from 'fastify';
 import supertest from 'supertest';
 
-import app from '../../app';
+import constructRoutes from '..';
 import {getAuthToken} from '../../util/authUtil';
 
 import type {FloodSettings} from '../../../shared/types/FloodSettings';
 
-const request = supertest(app);
+const app = fastify({disableRequestLogging: true, logger: false});
+let request: supertest.SuperTest<supertest.Test>;
+
+beforeAll(async () => {
+  await constructRoutes(app);
+  await app.ready();
+  request = supertest(app.server);
+});
+
+afterAll(async () => {
+  await app.close();
+});
 
 const authToken = `jwt=${getAuthToken('_config')}`;
