+++ server/routes/api/torrents.test.ts
@@ -1,14 +1,13 @@
-import axios from 'axios';
 import crypto from 'crypto';
+import fastify from 'fastify';
 import fs from 'fs';
-import MockAdapter from 'axios-mock-adapter';
 import os from 'os';
 import path from 'path';
 import readline from 'readline';
 import stream from 'stream';
 import supertest from 'supertest';
 
-import app from '../../app';
+import constructRoutes from '..';
 import {getAuthToken} from '../../util/authUtil';
 import {getTempPath} from '../../models/TemporaryStorage';
 import paths from '../../../shared/config/paths';
@@ -20,9 +19,20 @@ import type {TorrentList} from '../../../shared/types/Torrent';
 import type {TorrentStatus} from '../../../shared/constants/torrentStatusMap';
 import type {TorrentTracker} from '../../../shared/types/TorrentTracker';
 
-const request = supertest(app);
+const app = fastify({bodyLimit: 100 * 1024 * 1024 * 1024, disableRequestLogging: true, forceCloseConnections: true});
+let request: supertest.SuperTest<supertest.Test>;
 
+const activityStream = new stream.PassThrough();
 const authToken = `jwt=${getAuthToken('_config')}`;
+const rl = readline.createInterface({input: activityStream});
+
+beforeAll(async () => {
+  console.time('before all');
+  await constructRoutes(app);
+  await app.ready();
+  request = supertest(app.server);
+  request.get('/api/activity-stream').send().set('Cookie', [authToken]).pipe(activityStream);
+});
 
 const tempDirectory = getTempPath('download');
 
@@ -33,19 +43,9 @@ const torrentFiles = [
   path.join(paths.appSrc, 'fixtures/multi.torrent'),
 ].map((torrentPath) => Buffer.from(fs.readFileSync(torrentPath)).toString('base64')) as [string, ...string[]];
 
-const mock = new MockAdapter(axios, {onNoMatch: 'passthrough'});
-
-mock
-  .onGet('https://www.torrents/single.torrent')
-  .reply(200, fs.readFileSync(path.join(paths.appSrc, 'fixtures/single.torrent')));
-
-mock
-  .onGet('https://www.torrents/multi.torrent')
-  .reply(200, fs.readFileSync(path.join(paths.appSrc, 'fixtures/multi.torrent')));
-
 const torrentURLs: [string, ...string[]] = [
-  'https://www.torrents/single.torrent',
-  'https://www.torrents/multi.torrent',
+  'https://releases.ubuntu.com/focal/ubuntu-20.04.6-live-server-amd64.iso.torrent',
+  'https://flood.js.org/api/test-cookie',
 ];
 
 const torrentCookies = {
@@ -58,12 +58,8 @@ const testTrackers = [
   `http://${crypto.randomBytes(8).toString('hex')}.com/announce.php?key=test`,
 ];
 
-const torrentHashes: string[] = [];
-const createdTorrentHashes: string[] = [];
-
-const activityStream = new stream.PassThrough();
-const rl = readline.createInterface({input: activityStream});
-request.get('/api/activity-stream').send().set('Cookie', [authToken]).pipe(activityStream);
+let torrentHash = '';
+let createdTorrentHash = '';
 
 const watchTorrentList = (op: 'add' | 'remove' | 'replace' | 'move' | 'copy' | 'test'): Promise<void> => {
   return new Promise((resolve) => {
@@ -78,23 +74,6 @@ const watchTorrentList = (op: 'add' | 'remove' | 'replace' | 'move' | 'copy' | '
     });
   });
 };
-
-beforeAll((done) => {
-  request
-    .get('/api/client/connection-test')
-    .send()
-    .set('Cookie', [authToken])
-    .set('Accept', 'application/json')
-    .expect(200)
-    .expect('Content-Type', /json/)
-    .end((err, res) => {
-      if (err) done(err);
-
-      expect(res.body).toMatchObject({isConnected: true});
-
-      done();
-    });
-});
 
 describe('POST /api/torrents/add-urls', () => {
   const addTorrentByURLOptions: AddTorrentByURLOptions = {
@@ -199,36 +178,11 @@ describe('POST /api/torrents/add-urls', () => {
               : ['stopped', 'inactive'];
             expect(torrent.status).toEqual(expect.arrayContaining(expectedStatuses));
 
-            torrentHashes.push(torrent.hash);
+            torrentHash = torrent.hash;
           }),
         );
 
         done();
-      });
-  });
-});
-
-describe('POST /api/torrents/delete', () => {
-  const torrentDeleted = watchTorrentList('remove');
-  it('Deletes added torrents', (done) => {
-    request
-      .post('/api/torrents/delete')
-      .send({hashes: torrentHashes, deleteData: true})
-      .set('Cookie', [authToken])
-      .set('Accept', 'application/json')
-      .expect(200)
-      .expect('Content-Type', /json/)
-      .end((err, _res) => {
-        if (err) done(err);
-
-        Promise.race([torrentDeleted, new Promise((r) => setTimeout(r, 1000 * 15))])
-          .then(async () => {
-            // Wait a while
-            await new Promise((r) => setTimeout(r, 1000 * 3));
-          })
-          .then(() => {
-            done();
-          });
       });
   });
 });
@@ -415,7 +369,7 @@ describe('POST /api/torrents/create', () => {
 
         await Promise.all(
           addedTorrents.map(async (torrent) => {
-            createdTorrentHashes.push(torrent.hash);
+            createdTorrentHash = torrent.hash;
             expect(torrent.isPrivate).toBe(false);
 
             if (process.argv.includes('--trurl')) {
@@ -435,7 +389,7 @@ describe('POST /api/torrents/create', () => {
 describe('PATCH /api/torrents/trackers', () => {
   it('Sets single tracker', (done) => {
     const setTrackersOptions: SetTorrentsTrackersOptions = {
-      hashes: [torrentHashes[0]],
+      hashes: [torrentHash],
       trackers: [testTrackers[0]],
     };
 
@@ -455,7 +409,7 @@ describe('PATCH /api/torrents/trackers', () => {
 
   it('Sets multiple trackers', (done) => {
     const setTrackersOptions: SetTorrentsTrackersOptions = {
-      hashes: [torrentHashes[0]],
+      hashes: [torrentHash],
       trackers: testTrackers,
     };
 
@@ -475,7 +429,7 @@ describe('PATCH /api/torrents/trackers', () => {
 
   it('GET /api/torrents/{hash}/trackers', (done) => {
     request
-      .get(`/api/torrents/${torrentHashes[0]}/trackers`)
+      .get(`/api/torrents/${torrentHash}/trackers`)
       .send()
       .set('Cookie', [authToken])
       .set('Accept', 'application/json')
@@ -497,7 +451,7 @@ describe('PATCH /api/torrents/trackers', () => {
 describe('GET /api/torrents/{hash}/contents', () => {
   it('Gets contents of torrents', (done) => {
     request
-      .get(`/api/torrents/${torrentHashes[0]}/contents`)
+      .get(`/api/torrents/${torrentHash}/contents`)
       .send()
       .set('Cookie', [authToken])
       .set('Accept', 'application/json')
@@ -520,7 +474,7 @@ describe('POST /api/torrents/move', () => {
 
   it('Moves torrent', (done) => {
     const moveTorrentsOptions: MoveTorrentsOptions = {
-      hashes: [createdTorrentHashes[createdTorrentHashes.length - 1]],
+      hashes: [createdTorrentHash],
       destination: destDirectory,
       moveFiles: true,
       isBasePath: true,
@@ -558,7 +512,7 @@ describe('POST /api/torrents/move', () => {
 
         expect(res.body.torrents == null).toBe(false);
         const torrentList: TorrentList = res.body.torrents;
-        const torrent = torrentList[createdTorrentHashes[createdTorrentHashes.length - 1]];
+        const torrent = torrentList[createdTorrentHash];
 
         expect(torrent).not.toBe(null);
         expect(torrent.directory.startsWith(destDirectory)).toBe(true);
